// web/src/lib/openapi.d.ts
export interface components {
    schemas: {
// ...
        CreateConnectionRequest: {
            /** @enum {string} */
            kind: "repo" | "org";
            owner: string;
            name?: string | null;
            /** Format: int64 */
            token_id: number;
            /** Format: date-time */
            backfill_from?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}